use a functional approach to parsing credo output
