+++ src/test/suite/parser.test.ts
@@ -1,22 +1,25 @@
 import * as vscode from 'vscode';
 import { expect } from 'chai';
-import CredoParser from '../../CredoParser';
+import { parseCredoIssue, parseCredoOutput } from '../../parser';
 import { CredoIssue } from '../../CredoOutput';
 
-describe('CredoParser', () => {
+describe('Parse Credo Output', () => {
   context('#parseCredoIssue', () => {
     it('parses a credo issue correctly into a vscode diagnostic', () => {
-      const parsedDiagnostic = CredoParser.parseCredoIssue({
-        category: 'readability',
-        check: 'Credo.Check.Readability.ModuleDoc',
-        column: 11,
-        column_end: 32,
-        filename: 'lib/sample_web/telemetry.ex',
-        line_no: 1,
-        message: 'Modules should have a @moduledoc tag.',
-        priority: 1,
-        trigger: 'SampleWeb.Telemetry',
-      }, '');
+      const parsedDiagnostic = parseCredoIssue({
+        issue: {
+          category: 'readability',
+          check: 'Credo.Check.Readability.ModuleDoc',
+          column: 11,
+          column_end: 32,
+          filename: 'lib/sample_web/telemetry.ex',
+          line_no: 1,
+          message: 'Modules should have a @moduledoc tag.',
+          priority: 1,
+          trigger: 'SampleWeb.Telemetry',
+        },
+        documentContent: '',
+      });
 
       expect(parsedDiagnostic.message).to.equal(
         'Modules should have a @moduledoc tag. (readability:Credo.Check.Readability.ModuleDoc)',
@@ -42,10 +45,10 @@ describe('CredoParser', () => {
       };
 
       it('marks the substring of the line when a trigger is given', () => {
-        const parsedDiagnostic = CredoParser.parseCredoIssue(
-          credoIssue,
-          'defmodule TestModule do\n  # TODO: any\nend\n',
-        );
+        const parsedDiagnostic = parseCredoIssue({
+          issue: credoIssue,
+          documentContent: 'defmodule TestModule do\n  # TODO: any\nend\n',
+        });
 
         expect(parsedDiagnostic.message).to.equal(
           'Found a TODO tag in a comment: # TODO: any. (design:Credo.Check.Design.TagTODO)',
@@ -62,10 +65,10 @@ describe('CredoParser', () => {
           ...credoIssue,
           trigger: null,
         };
-        const parsedDiagnostic = CredoParser.parseCredoIssue(
-          otherCredoIssue,
-          'defmodule TestModule do\n  # TODO: any\nend\n',
-        );
+        const parsedDiagnostic = parseCredoIssue({
+          issue: otherCredoIssue,
+          documentContent: 'defmodule TestModule do\n  # TODO: any\nend\n',
+        });
 
         expect(parsedDiagnostic.message).to.equal(
           'Found a TODO tag in a comment: # TODO: any. (design:Credo.Check.Design.TagTODO)',
@@ -81,32 +84,35 @@ describe('CredoParser', () => {
 
   context('#parseCredoOutput', () => {
     it('parses the credo output to a vscode diagnostic array', () => {
-      const parsedCollection = CredoParser.parseCredoOutput({
-        issues: [
-          {
-            category: 'readability',
-            check: 'Credo.Check.Readability.ModuleDoc',
-            column: 11,
-            column_end: 32,
-            filename: 'lib/sample_web/telemetry.ex',
-            line_no: 1,
-            message: 'Modules should have a @moduledoc tag.',
-            priority: 1,
-            trigger: 'SampleWeb.Telemetry',
-          },
-          {
-            category: 'design',
-            check: 'Credo.Check.Design.AliasUsage',
-            column: 7,
-            column_end: 32,
-            filename: 'test/support/channel_case.ex',
-            line_no: 35,
-            message: 'Nested modules could be aliased at the top of the invoking module.',
-            priority: -9,
-            trigger: 'Ecto.Adapters.SQL.Sandbox',
-          },
-        ],
-      }, { getText: () => '' } as vscode.TextDocument);
+      const parsedCollection = parseCredoOutput({
+        credoOutput: {
+          issues: [
+            {
+              category: 'readability',
+              check: 'Credo.Check.Readability.ModuleDoc',
+              column: 11,
+              column_end: 32,
+              filename: 'lib/sample_web/telemetry.ex',
+              line_no: 1,
+              message: 'Modules should have a @moduledoc tag.',
+              priority: 1,
+              trigger: 'SampleWeb.Telemetry',
+            },
+            {
+              category: 'design',
+              check: 'Credo.Check.Design.AliasUsage',
+              column: 7,
+              column_end: 32,
+              filename: 'test/support/channel_case.ex',
+              line_no: 35,
+              message: 'Nested modules could be aliased at the top of the invoking module.',
+              priority: -9,
+              trigger: 'Ecto.Adapters.SQL.Sandbox',
+            },
+          ],
+        },
+        document: { getText: () => '' } as vscode.TextDocument,
+      });
 
       expect(parsedCollection.length).to.equal(2);
     });
